requirements.txt
torch==2.5.1
torchvision
#####如果你需要显卡加速，请安装对应版本的CUDA，例如torch==2.5.1+cu121#####
######在install_pytorch_cpu&gpu.txt中有详细说明#####
numpy<2
opencv-python
timm
scipy
scikit-image
kornia
einops

tqdm
prettytable
tabulate

huggingface-hub>0.25
accelerate
gradio
moviepy
Pillow
